currency_monitor/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-simple-history==3.3.0
django-timezone-field==5.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
